god_brain_core/requirements.txt
openai>=1.0.0
faiss-cpu
numpy
python-dotenv
loguru
httpx
nats-py
